RA-92: Changed output data from Double to Integer to maintain consistency (#152)
* RA-92: Changed output data from Double to Integer to maintain consistency

* refactor: added comment to explain why to divide by two to geht the TxSize in Bytes

* RA-92: Resolved merge conflict

---------

+++ api/src/main/java/org/cardanofoundation/rosetta/api/construction/service/CardanoConstructionService.java
@@ -35,14 +35,14 @@ public interface CardanoConstructionService {
   TransactionParsed parseTransaction(NetworkIdentifierType networkIdentifierType,
       String transaction, boolean signed);
 
-  Double checkOrReturnDefaultTtl(Integer relativeTtl);
+  Integer checkOrReturnDefaultTtl(Integer relativeTtl);
   Long updateTxSize(Long previousTxSize, Long previousTtl, Long updatedTtl);
   Long calculateTxMinimumFee(Long transactionSize, ProtocolParams protocolParameters);
 
   Signatures signatureProcessor(EraAddressType eraAddressType, AddressType addressType,
       String address);
 
-  Double calculateTxSize(NetworkIdentifierType networkIdentifierType, List<Operation> operations, int ttl, DepositParameters depositParameters);
+  Integer calculateTxSize(NetworkIdentifierType networkIdentifierType, List<Operation> operations, int ttl, DepositParameters depositParameters);
 
   String buildTransaction(String unsignedTransaction,
       List<Signatures> signaturesList, String transactionMetadata);
